fix stim relabeling
Was previously indexing columns by the mark number, but that's a problem when the trigger/mark IDs don't start at 1. Now the index of a given name/mark in the array of Conditions is matched to the column index of OnsetMatrix.

There was a missing case for when we want to replace an entire condition (e.g., all m2), but there's only one instance of that trigger. MCP_relabel_stimuli was defaulting to the reorganizing behavior intended for a list of different stimulus labels. Added a new case to handle the 1-to-1 match more simply (swap the Condition.Name field without making any other change).